worker/requirements.txt
praw>=7.7.0
anthropic>=0.39.0
python-dotenv>=1.0.0
requests>=2.31.0
